requirements.txt
cachetools==3.1.1
cycler==0.10.0
decorator==4.4.0
Django==2.2.1
feedgen==0.7.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.12.3
imageio==2.5.0
kiwisolver==1.1.0
lxml==4.3.3
matplotlib==3.1.0
networkx==2.3
numpy==1.16.3
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
rsa==4.0
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
youtube-dl==2019.5.20